Added AI algorithm function to make the next move.
It's a hearth of the game, so it's a good place to start. I think the function could be improved by adding a context of possible failure, but I'm not sure I need that yet.

Currently the function works on assumption that the player input has been validated and sanitized.

When it comes to algorithm, it's something that comes out of doing math on paper - if you want to force the other player to take the last coin, you have to keep him multiples of 4 away from it, so there would be nothing he can do. (because he can take up to 3 coins).

There's this sequence of numbers, if you started your turn on it - there's nothing meaningful you can do, so I hardcoded '1' to save myself extra code complication. If you didn't start on it, you make sure the other player does.