Added turn action function.
Not sure if this is the best way to do it. Looking at this function screams of refactoring. But I don't know enough on how to work with Haskells IO, and what's a good standard. The thing is I need those logs!

In short, this function takes a number of coins on the table and the player who's turn now has to be as its arguments.

Based of player variable it decides on whether to call an AI code or the Player code and get his input. Both cases it returns a new number of coins as it's output (wrapped up in IO context).

Guess I could refactor AI do block and Human do block into their own functions at some point.