Added the game setup - startTheGame function.
It essentially initializes the game state for the 'turn' function, which is the game itself.

Here the random number of 'coins' is generated. For the sake of it being easy to test the game, I kept the max range number low (33). This is very boring game, trust me, you wouldn't want to sit through billion of coins. Also the player is asked if he wants to go first - this was a requirement. The algorithm has to be tested in 3 games of AI vs Human, and 3 games of Human vs AI.

Added the main block, which is the entry point for the app.

By this point application has to be compiled to be tested.